small bugfixes and warning fixes

+++ src/utils/Math.hpp
@@ -33,6 +33,6 @@ inline float ease_out_back(float progress) {
 inline float smoothstep(float x, float y, float progress) { return ccm::ext::smoothstep(x, y, progress); }
 inline bool same_parity(float a, float b) { return ((static_cast<int>(a) ^ static_cast<int>(b)) & 1) == 0; }
 inline bool same_sign(float a, float b) { return a * b >= 0.f; }
-inline uint8_t get_uint8_from_normal(float normal) { return std::lerp(0, 255, normal); }
+inline uint8_t get_uint8_from_normal(float normal) { return static_cast<uint8_t>(std::lerp(0, 255, normal)); }
 
 } // namespace util
